:bug: fix(vercel): prevent prerender error by moving subscription checkout client logic into Client component and wrapping page in Server Component + Suspense (Next.js 15 pattern)
